incorporating image upload component into new dog form
